Refactor refatoring class and add documentation
