+++ main_binary_classification.py
@@ -1,15 +1,15 @@
-from configs import DATASETS
+from configs import DATASETS, Level
 from db.DBConnector import close_connection
 from ml.models.builder import build_models
 from ml.pipelines.binary import BinaryClassificationPipeline
-from refactoring import build_refactorings
+from ml.refactoring import build_refactorings
 from utils.log import log_init, log_close, log
 import datetime
 
 log_init(f"results/classifier_training_{datetime.datetime.now().strftime('%Y-%m-%d-%H-%M-%S')}.txt")
 log("ML4Refactoring: Binary classification")
 
-refactorings = build_refactorings()
+refactorings = build_refactorings(Level)
 
 # Run models
 models = build_models()
